:children_crossing: Added removal confirmation

+++ src/commands/remove/session.ts
@@ -1,5 +1,6 @@
 import {Command, flags} from '@oclif/command'
 import {existsSync, readJSON, writeFile} from 'fs-extra'
+import {prompt} from 'inquirer'
 
 import {getDayFile} from '../../get-day-file'
 
@@ -8,12 +9,13 @@ export default class RemoveSession extends Command {
 
   static flags = {
     help: flags.help({char: 'h'}),
+    confirm: flags.boolean({char: 'c'})
   }
 
   static args = [{name: 'id', description: 'ID of session to remove\nfind with: tim sessions -x', required: true}]
 
   async run() {
-    const {args} = this.parse(RemoveSession)
+    const {args, flags} = this.parse(RemoveSession)
     let id = args.id
 
     const now = new Date()
@@ -23,12 +25,21 @@ export default class RemoveSession extends Command {
       const filteredSessions = day.sessions.filter(({id: sessionID}: {id: string}) => sessionID !== id)
 
       if (filteredSessions.length !== day.sessions.length) {
-        day.sessions = filteredSessions
+        let confirm = {confirm: false}
+        if (!flags.confirm) {
+          confirm = await prompt({type: 'confirm', name: 'confirm', message: `Are you sure you want to remove: ${id}?`})
+        }
 
-        const json = JSON.stringify(day, null, 4)
-        await writeFile(dayFile, json, 'utf-8')
+        if (flags.confirm || confirm.confirm) {
+          day.sessions = filteredSessions
 
-        this.log(`Removed session with ID: ${id}`)
+          const json = JSON.stringify(day, null, 4)
+          await writeFile(dayFile, json, 'utf-8')
+
+          this.log(`Removed session with ID: ${id}`)
+        } else {
+          this.log(`Canceled removal of session: ${id}`)
+        }
       } else {
         this.log(`No session exists with ID: ${id}`)
       }
